requirements.txt
boto3==1.35.44
fastapi==0.108.0
uvicorn==0.30.6
pydantic==2.8.2
requests==2.32.3
pillow==9.5.0
python-multipart==0.0.9
websocket
